Attack state improvements, Vaulting Component, Combat Camera
Made improvements to the jump attack state, final state of the attack animation now only triggers when the player is falling and is x amount of units from the ground instead of firing when the player is grounded.

Implemented vaulting component and climbing state allowing the player to climb over obstacles. Still need to implement animations but ledge detection and movement work well enough.

Combat camera implemented, zooming the camera out based on how many enemies are currently engaged in combat with the player. Implementation not final will be changing how enemies are considered to be engaged in combat.